requirements.txt
anyio==3.6.1
asgiref==3.5.2
boto3==1.23.9
botocore==1.26.10
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django
et-xmlfile==1.1.0
fastapi==0.78.0
Flask==2.1.2
h11==0.13.0
idna==3.3
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
numpy
opencv-python==4.1.2.30
openpyxl==3.0.10
pandas
Pillow==9.1.1
pydantic==1.9.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.2
workbook==1.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlutils==2.0.0
xlwt==1.3.0
